RaspiPICprog documentation for PIC programming interface PCB production.
Example PCB for programming "hobby" PIC processors with Raspberry Pi. The PCB
is produced with toner transfer method. An external 12 V regulated power
supply is required - usually available in ham shack, but a SLAB or 8xAA
batteries will do the job too. The interface can also be used as a low power
+5 V regulated power supply when testing the PIC.