Do not reinitialize GPIO, T1CON and WPU after WDT reset.
This should make operation after false interrupt detection on GP2/INT more
correct. There will be still short time when all digital I/O's are set as
inputs because WDT reset sets all bits in TRISIO to one. If the power switch
or power supply is slow enough this should not be a problem.