Flash LED on GP5 when timed command active.
If the most significant bit in timer1 is one use 'sleep' instruction to save
power. Not sure if this is really useful. Seems to have side effect of slower
I2C reply from the PIC. The normal operation can be revovered using timer
reset command 0x50 since this will also reset the MSB in timer1.