First version of 'pipipowerd.c' for monitoring and controlling power supply.
The 'pipipowerd.c' tests first i2c data flow to the power supply controlling
PIC. Then the PIC internal timer is read. In the main loop the battery
voltage is read and preprogrammed intervals. If the voltage is too low
automatic shutdown and power off is done. If the button is pressed the
red LED is lit to indicate that the power supply is waiting for confirmation
for the shutdown. If the button is pressed again, the system is shutdown
and the power supply is switched off after preset time interval. The operation
can be checked from the log file 'pipicpowerd.log'.