pipicpowerd.c stop button/volt read if shutdown imminent
The 'pwroff' is initially 0 and set to 1 or 2 if system is going to
halt state with command 'shutdown -h now'. If 1 the PIC is programmed
to power off after a preset time delay and then power up after a second
usually much longer time delay. If 2 only delayed power down is
programmed to the PIC.

This should fix the bug that set the red LED on just before the power
down.