pipicpowerd.c option to reset PIC timer if i2c test fails
In the configuration file if FORCERESET is set the command 0x50 is
sent to the PIC to reset the internal timer in case of failure in
sending random 32 bits and reading them back. This should allow
automatic recovery in case that PIC was power cycled, but is
not correct behaviour if there is a cable problem to the PIC. Thus
this option should be used with care.