Add forced power off to pipicpowerd.c
If FORCEPOWEROFF is set in the configuration file the power plug is
abruptly pulled out after the set PIC cycle counts. Normally before
powering down the system is shutdown cleanly, but this does not happen
in case of forced power down. Thus the forced power down will lead
to data loss and might cause data corruption on the SD card and other
pheripherals.

If the system needs to reboot after pulling out the power plug, the
FORCEPOWERON can be used. The power plug is put in again after the
set PIC cycles.