pipicpowerd.c estimate time left before battery empty
The remaining battery capacity is estimated from the measured voltage.
Temperature corrected voltage needs to be used if the power supply is outside.
From the estimated current and nominal battery capacity the remaining time
before the battery is empty is calculated. Temperature independent
Peukert's law is used for this. Note that for long battery life it is not
good idea to run the lead acid battery completely empty.