pipicpowerd.c power down on low battery charge
If the battery charge is lower than the value set by MINBATTLEVEL in
configuration file, system shutdown is done in five minutes. During this
time no new logins are allowed. After the system shutdown the power is
switched off with the delay set by PWRDOWN in the configuration file.
The file /var/lib/pipicpowerd/wakeup can be used to automatically wakeup
the system in future.

If the battery voltage is higher than MAXBATTVOLTS in configuration file
a warning message is send to all the users. The charging voltage should be
reduced if that is possible for example by moving the solar panel to shadow
from direct sun light.

The remaining battery capacity is now more correctly estimated from the
measured loaded battery voltage.