pipicpowerd.c check at start if ntpd is running
In function test_ntp() system command 'ntpq -p' is called and the output
is written to a temporary file. This temporary file is read to see if
the 'ntpd' seems to be running. If this is the case the PIC counter is
not used to set the system time at power up.