pipicpowerd.c add battery monitoring scripts
The commands from file 'battery.sh' can be added to the end of .bashrc
to see the battery status during remote login.

The 'filterlog' can be used to extract battery data from the pipicpowerd.log
file. These values can then be plotted with gnuplot, see MONITORBATTERY.txt
for an example.