pipicpowerd.c do not reset PIC counter at power down any more
The PIC counter is not reset at power down any more. Instead the value
is saved to file /var/lib/pipicpowerd/timer. This file is read at boot
time to estimate system time if the ntpd can not get time from anywhere
else.

Keeping the PIC counter running helps to do sanity checks on the system
time and debug the code. In the configuration file COUNTINT can be set
to log PIC counter values to the log file.