pipicpowerd.c more careful power down and power up handling
Before 'shutdown' command the 'sync' is now executed to write data
on the disk or SD card.

The /var/lib/pipicpowerd/pwrup file is used to tell to the daemon if
the start is due to power up or simple restart of the service
from command line. This file is now handeled (hopefully) more carefully.