pipicpowerd.c add WiFi monitoring and auto power cycle options
The WIFINT in the configuration file can be used to set time interval
for reading the WiFi operation status usually in
/sys/class/net/wlan0/operstate . If the WiFi is down and WIFIACT is
set to one a system reboot is done in hope that WiFi will be back after
reboot. If the WIFIACT is set to two, the Raspberry Pi is power
cycled hoping that the WiFi will start to work after rebooting.
If WIFIACT is one or two the WIFINT should not be too short since this
could result in permanent system rebooting or power cycling.