pipicpowerd.c cyclic power up operation based on PIC timer
If the file /var/lib/pipicpowerd/puptime exists and has integer value
of minutes the system is shutdown followed by power down after the
given minutes. The power will come back after the minutes given in
file /var/lib/pipicpowerd/pdowntime. The timing is based on PIC
counter and does not require that the system time is set correctly
after the boot up.