pipicpowerd.c option for small solar panel operation
In cycling operation given by files /var/log/pipicpowerd/puptime and
/var/log/pipicpowerd/pdowntime in case that the battery is already fully
charged the power down is not done if SOLARPOWER is set in configuration
file. This should allow to consume part of the energy that can not be
stored to the battery.

Modify i2c error handling in functions powerdown() and terminate().