pipicpowerd.c automatic update of power uptime in cyclic operation
The /var/log/pipicpowers.log file can be used to calculate average
uptime for last days. The number of days is set in the configuration
file with SOLARDAYS parameter. The SOLARCYCLE is the total minutes
in cyclic power up - power down operation. The files
/var/lib/pipicpowerd/puptime and /var/lib/pipicpowerd/pdowntime
are set accordingly to use the battery in such a way that maximum
computer time is available without over filling the battery nor
running it to very low charge state.